+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changed
 - Updated schug library to v1.10
 - Disable SQLAlchemy logger
+- To avoid timeout errors, update genes, transcripts exons only from pre-downloaded files from schug
 
 ## [3.1]
 ### Changed